+++ foth/eval/builtins_test.go
@@ -1,4 +1,4 @@
-package main
+package eval
 
 import (
 	"testing"
